+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * rename `ModelManager.find_all` to `ModelManager.all` and `ModelManager.find_by` to `ModelManager.filter_by` for consistency with the `Query` api
 * move instantiation of the `CSRFProtect` extension from the security bundle into the controller bundle, where it belongs, so that it always gets used
 * improve registration of request cycle functions meant to run only for a specific bundle blueprint
+* update `BaseService` to use a MetaOptionsFactory
 
 ## 0.5.1 (2018/07/25)